+++ server/src/main/java/org/allaymc/server/entity/component/EntityBaseComponentImpl.java
@@ -226,20 +226,10 @@ public class EntityBaseComponentImpl implements EntityBaseComponent {
         var dimension = getDimension();
         dimension.forEachBlockStates(aabb, 0, (x, y, z, blockState) -> {
             var block = new BlockStateWithPos(blockState, new Position3i(x, y, z, dimension), 0);
-            var entityHasBlockCollision = this.hasBlockCollision();
-            var blockCanCollideWithEntity = blockState.getBehavior().canCollideWithEntity();
 
-            // When either of them can collide with the other, we need to check for collisions
-            if (entityHasBlockCollision || blockCanCollideWithEntity) {
-                if (blockState.getBlockStateData().collisionShape().translate(x, y, z).intersectsAABB(aabb)) {
-                    if (entityHasBlockCollision) {
-                        this.onCollideWithBlock(block);
-                    }
-
-                    if (blockCanCollideWithEntity) {
-                        blockState.getBehavior().onCollideWithEntity(block, thisEntity);
-                    }
-                }
+            if (blockState.getBlockStateData().collisionShape().translate(x, y, z).intersectsAABB(aabb)) {
+                this.onCollideWithBlock(block);
+                blockState.getBehavior().onCollideWithEntity(block, thisEntity);
             }
 
             // Check if the entity is inside block