+++ server/src/main/java/org/allaymc/server/entity/component/projectile/EntityEnderPearlBaseComponentImpl.java
@@ -19,8 +19,8 @@ public class EntityEnderPearlBaseComponentImpl extends EntityProjectileBaseCompo
     }
 
     @Override
-    public void onCollideWithEntity(Entity other) {
-        if (this.willBeDespawnedNextTick() || other == shootingEntity || !callProjectileHitEvent(null)) {
+    protected void onHitEntity(Entity other) {
+        if (this.willBeDespawnedNextTick() || other == shootingEntity) {
             return;
         }
 
@@ -33,8 +33,8 @@ public class EntityEnderPearlBaseComponentImpl extends EntityProjectileBaseCompo
     }
 
     @Override
-    public void onCollideWithBlock(BlockStateWithPos block) {
-        if (this.willBeDespawnedNextTick() || !callProjectileHitEvent(block)) {
+    protected void onHitBlock(BlockStateWithPos block) {
+        if (this.willBeDespawnedNextTick()) {
             return;
         }