+++ server/src/main/java/org/allaymc/server/entity/component/projectile/EntitySplashPotionBaseComponentImpl.java
@@ -34,8 +34,8 @@ public class EntitySplashPotionBaseComponentImpl extends EntityProjectileBaseCom
     }
 
     @Override
-    public void onCollideWithEntity(Entity other) {
-        if (this.willBeDespawnedNextTick() || other == shootingEntity || !callProjectileHitEvent(null)) {
+    protected void onHitEntity(Entity other) {
+        if (this.willBeDespawnedNextTick() || other == shootingEntity) {
             return;
         }
 
@@ -44,8 +44,8 @@ public class EntitySplashPotionBaseComponentImpl extends EntityProjectileBaseCom
     }
 
     @Override
-    public void onCollideWithBlock(BlockStateWithPos block) {
-        if (this.willBeDespawnedNextTick() || !callProjectileHitEvent(block)) {
+    protected void onHitBlock(BlockStateWithPos block) {
+        if (this.willBeDespawnedNextTick()) {
             return;
         }